[busybox] Fix telnetd install before inetd As telnetd doesn't depend on inetd it's not safe to assume inetd.conf already exists before inst

+++ meta-oe/recipes-core/busybox/busybox_1.23.1.bbappend
@@ -55,6 +55,8 @@ pkg_postinst_${PN}_append () {
 }
 
 pkg_preinst_${PN}-telnetd_prepend () {
+if [ -e $D/etc/inetd.conf ]; then
 	grep -vE '^#*\s*(23|telnet)' $D/etc/inetd.conf > $D/tmp/inetd.tmp
 	mv $D/tmp/inetd.tmp $D/etc/inetd.conf
+fi
 }
